+++ src/values_list.rs
@@ -2,11 +2,23 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-use {Stream, Length, Error};
+use {Stream, TextFrame, Length, Error};
 
 /// Iterator over a list of \<number\> types.
 #[derive(Copy,Clone,PartialEq)]
-pub struct NumberList<'a>(pub Stream<'a>);
+pub struct NumberList<'a>(Stream<'a>);
+
+impl<'a> NumberList<'a> {
+    /// Constructs a new `NumberList` from `TextFrame`.
+    pub fn from_frame(frame: TextFrame<'a>) -> NumberList<'a> {
+        NumberList(Stream::from_frame(frame))
+    }
+
+    /// Returns an underling string.
+    pub fn data(&self) -> &str {
+        self.0.slice()
+    }
+}
 
 impl<'a> Iterator for NumberList<'a> {
     type Item = Result<f64, Error>;
@@ -22,7 +34,19 @@ impl<'a> Iterator for NumberList<'a> {
 
 /// Iterator over a list of \<length\> types.
 #[derive(Copy,Clone,PartialEq)]
-pub struct LengthList<'a>(pub Stream<'a>);
+pub struct LengthList<'a>(Stream<'a>);
+
+impl<'a> LengthList<'a> {
+    /// Constructs a new `LengthList` from `TextFrame`.
+    pub fn from_frame(frame: TextFrame<'a>) -> LengthList<'a> {
+        LengthList(Stream::from_frame(frame))
+    }
+
+    /// Returns an underling string.
+    pub fn data(&self) -> &str {
+        self.0.slice()
+    }
+}
 
 impl<'a> Iterator for LengthList<'a> {
     type Item = Result<Length, Error>;